A scratch refactoring on making battles more modular
And the code supporting them clearer.

In its current state this is probably a step down
